feat: update bug in async table creation

+++ src/tradingtoolbox/clickhouse/clickhouse_async.py
@@ -64,12 +64,15 @@ class ClickhouseAsync(BaseModel):
             table_name: The name of the table to insert into
             drop: Whether to drop the table if it already exists
         """
+        if len(df) == 0:
+            return
+
         if drop:
             await self.drop_table(table_name)
         schema = generate_table_schema(df, table_name)
 
-        await self.async_client.insert_df(table_name, df)
         await self.async_client.command(schema)
+        await self.async_client.insert_df(table_name, df)
 
     async def execute_command(self, query: str):
         """
